+++ app/Http/Controllers/Admin/TableController.php
@@ -4,13 +4,18 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\TableStoreRequest;
+use App\Http\Service\TableService;
 use App\Models\Table;
-use Illuminate\Http\Request;
-
-use function Ramsey\Uuid\v1;
 
 class TableController extends Controller
 {
+    private $tableService;
+
+    public function __construct(TableService $tableService)
+    {
+        $this->tableService = $tableService;
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -29,7 +34,8 @@ class TableController extends Controller
      */
     public function create()
     {
-        return view('admin.tables.create');
+        $tableLocations = $this->tableService->getTableLocations();
+        return view('admin.tables.create', compact('tableLocations'));
     }
 
     /**
@@ -40,11 +46,12 @@ class TableController extends Controller
      */
     public function store(TableStoreRequest $request)
     {
+
         Table::create([
             'name' => $request->name,
             'guest_number' => $request->guest_number,
             'status' => $request->status,
-            'location' => $request->location,
+            'table_location' => $request->location,
         ]);
 
         return to_route('admin.tables.index')->with('success', 'Table created successfully.');
@@ -95,6 +102,12 @@ class TableController extends Controller
     public function destroy(Table $table)
     {
         $table->reservations()->delete();
+
+        // check if table has orders
+        if ($table->orders()->count() > 0) {
+            return to_route('admin.tables.index')->with('danger', 'Table has orders, cannot delete.');
+        }
+
         $table->delete();
 
         return to_route('admin.tables.index')->with('danger', 'Table daleted successfully.');
